fixed to get swift version before concurrent parsing
